refactor: formatDateTime を date にまとめる
